+++ test/integration/runner/features/samples/TestFeatureEmptyLabel.php
@@ -17,7 +17,7 @@
  * Copyright (c) 2016 (original work) Open Assessment Technologies SA;
  *
  */
-namespace oat\taoTests\test\runner\features\samples;
+namespace oat\taoTests\test\integration\runner\features\samples;
 
 use oat\taoTests\models\runner\features\TestRunnerFeature;
 
@@ -26,11 +26,11 @@ use oat\taoTests\models\runner\features\TestRunnerFeature;
  *
  * @author Christophe Noël <christophe@taotesting.com>
  */
-class TestFeature extends TestRunnerFeature
+class TestFeatureEmptyLabel extends TestRunnerFeature
 {
     public function getLabel()
     {
-        return __('testFeature');
+        return '';
     }
 
     public function getDescription()
